Add support for launching a public class
If the first class declared in the initial file
does not have a main entry point, check if the
file has a public top-level class with a name
matching the name of the file and with a main
entry point. If there is one, invoke the main
entry point of that class, otherwise fail.